pyproject.toml
[project]
name = "new-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.124.4",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "cloudinary>=1.41.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
